+++ src/services/business/barcode/Shopfloor.Barcode.Application/Command/Imports/DeleteImportCommand.cs
@@ -1,6 +1,5 @@
 using MediatR;
 using Shopfloor.Barcode.Domain.Interfaces;
-using Shopfloor.Core.Exceptions;
 using Shopfloor.Core.Models.Responses;
 
 namespace Shopfloor.Barcode.Application.Command.Imports
@@ -19,7 +18,8 @@ namespace Shopfloor.Barcode.Application.Command.Imports
 
         public async Task<Response<int>> Handle(DeleteImportCommand command, CancellationToken cancellationToken)
         {
-            var entity = await _repository.GetImportByIdAsync(command.Id) ?? throw new ApiException($"Import Not Found (Id:{command.Id}).");
+            var entity = await _repository.GetImportByIdAsync(command.Id);
+            if (entity == null) return new($"Import Not Found (Id:{command.Id}).");
             var articleBarcodes = entity.ImportArticles.SelectMany(x => x.ImportDetails).Select(y => y.ArticleBarcode);
             await _repository.DeleteImportAsync(entity, articleBarcodes);
             return new Response<int>(entity.Id);
